refactor database models and create migration scripts
